Add comments explaining `RoomLandscapeMaskSprite`
